Rename test classes and put tests into top level classes.
JENA-624
AbstractTest* for the abstract test classes.
TS_ for the test suites.
Collect each TS_* into test suite collection TC_General
