+++ jena-arq/src/test/java/org/apache/jena/sparql/core/mem/AbstractTestQuadTable.java
@@ -18,39 +18,33 @@
 
 package org.apache.jena.sparql.core.mem;
 
-import static org.apache.jena.graph.Node.ANY;
+import static java.util.EnumSet.allOf;
+import static org.apache.jena.ext.com.google.common.collect.Sets.powerSet;
 import static org.apache.jena.graph.NodeFactory.createURI;
-import static org.apache.jena.sparql.core.mem.QuadTableTest.quadQueryPatterns;
-import static org.apache.jena.sparql.core.mem.TupleSlot.GRAPH;
 
 import java.util.Set;
 import java.util.stream.Stream;
 
 import org.apache.jena.graph.Node;
-import org.apache.jena.graph.Triple;
+import org.apache.jena.sparql.core.Quad;
 
-public abstract class TripleTableTest extends TupleTableTest<Triple, TripleTable> {
+public abstract class AbstractTestQuadTable extends AbstractTestTupleTable<Quad, QuadTable> {
 
-	private static final Node sampleNode = createURI("info:test");
-
-	private static final Triple testTriple = Triple.create(sampleNode, sampleNode, sampleNode);
-
-	@Override
-	protected Triple testTuple() {
-		return testTriple;
-	}
+	protected static final Node sampleNode = createURI("info:test");
+	private static final Quad q = Quad.create(sampleNode, sampleNode, sampleNode, sampleNode);
 
 	@Override
-	protected Stream<Triple> tuples() {
-		return table().find(ANY, ANY, ANY);
+	protected Quad testTuple() {
+		return q;
 	}
 
 	@Override
 	public Stream<Set<TupleSlot>> queryPatterns() {
-		return tripleQueryPatterns();
+		return quadQueryPatterns();
 	}
 
-	static Stream<Set<TupleSlot>> tripleQueryPatterns() {
-		return quadQueryPatterns().filter(s -> !s.contains(GRAPH));
+	static Stream<Set<TupleSlot>> quadQueryPatterns() {
+		return powerSet(allOf(TupleSlot.class)).stream();
 	}
+
 }
