+++ jena-arq/src/test/java/org/apache/jena/sparql/core/mem/AbstractTestTupleTableForms.java
@@ -19,7 +19,7 @@
 package org.apache.jena.sparql.core.mem;
 
 import static java.util.stream.Collectors.toSet;
-import static org.apache.jena.sparql.core.mem.TupleTableTest.allWildcardQuery;
+import static org.apache.jena.sparql.core.mem.AbstractTestTupleTable.allWildcardQuery;
 
 import java.util.Map;
 import java.util.Set;
@@ -29,7 +29,7 @@ import java.util.stream.Stream;
 import org.junit.Assert;
 import org.junit.Test;
 
-public abstract class TupleTableFormsTest<TableForm extends Predicate<Set<TupleSlot>>> extends Assert {
+public abstract class AbstractTestTupleTableForms<TableForm extends Predicate<Set<TupleSlot>>> extends Assert {
 
 	protected abstract Stream<TableForm> tableForms();
